moved main function
